Particle stuff

+++ src/main/java/mod/schnappdragon/habitat/client/particle/FairyRingSporeParticle.java
@@ -22,7 +22,7 @@ public class FairyRingSporeParticle extends TextureSheetParticle {
         this.xo = this.x;
         this.yo = this.y;
         this.zo = this.z;
-        if (this.age++ >= this.lifetime || this.onGround || this.level.isFluidAtPosition(new BlockPos(this.x, this.y, this.z), (fluidState) -> !fluidState.isEmpty()))
+        if (this.age++ >= this.lifetime || this.onGround || this.level.isFluidAtPosition(new BlockPos((int) this.x, (int) this.y, (int) this.z), (fluidState) -> !fluidState.isEmpty()))
             this.remove();
         else {
             this.yd -= this.gravity;
